Updated build.gradle
This change was applied to all the examples after they were imported using WPILib. If it is not applied, the build always fails but you can continue it. Adding the version of this dependency allows the build to succeed without pausing for input.

Other changes are incidental and not important